Disable branching with disable-branching property.
Provide developer the ability to disable branching with disable-branching = true parameter.
When branching is not available or we do not want to use it, and we are developing in feature branch plugin fail to download translations. That is because it's using branch name on API request. With that change we ignore that parameter in API request and translation download completes successfully.